loop write buffers until showtdown is requested
if we would like to shutdown the client we are blocked until writeBuffers is completed, but it could be that for example the server is down or client are offline. In this case the while (true) loop is never finished and shutdownSemaphore is never released and so you can not shutdown the client.