don't run into exception when you are offline on startup
If you start the InfluxDBManager but you are offline you run into an UnknownHostException, we catch this, give an error log but don't throw the error. In this situation when you go online later you are able to write previously collected points.
In this scenario you can go on and offline and so on without exceptions.